requirements.txt
streamlit>=1.32.0
snowflake-snowpark-python>=1.5.1
pandas>=2.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
